refactor: extract admin status tracking actions
